Header & Implementation
Added a copy constructor and a current_index.
Changed the name of the dynamic array to "symptoms"

Included the implementation for all the functions for class Patient.  

******** WHAT NEEDS TO BE ADDED**********
prototypes and implementations for forward chaining and backwards chaining need to be written. The backward chaining function definitions are commented at the top of the header file. The definitions are my translations and my be incorrect.